Settings

+++ GoodReads/src/main/java/com/example/goodreads/model/entities/Book.java
@@ -9,7 +9,7 @@ import java.util.Set;
 public class Book {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int bookId;
+    private long bookId;
 
     @Column
     private String title;
